add restrictions and styles
